requirements.txt
streamlit==1.45.1
python-dotenv==1.1.0
langchain==0.3.25
langchain-google-genai==2.1.5
langchain-chroma==0.2.4
langchain-text-splitters==0.3.8
chromadb==1.0.10
SQLAlchemy==2.0.41
pandas==2.2.3
numpy==2.2.6
pydantic==2.11.5
protobuf==5.29.4
tiktoken==0.9.0
aiohttp==3.12.1
aiosignal==1.3.2
aiosqlite==0.21.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0rc1
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
marshmallow==3.26.1
multidict==6.4.4
orjson==3.10.18
packaging==24.2
pyarrow==20.0.0
pycparser==2.22
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
langchain_community